Moving generation machinery to the instruments
